Megacommit, did every fuckin thing. See description.
Added a database for users, they are no longer stored in the configuration.
Added a frontend fo registration, signing in, obtaining an access token.
Added an endpoint that serves the Presenti renderer that points to a given scope. There's an option for disabling CSS.
Rewrote the Remote Presence API to efficiently handle a large number of users.
Added commands to the Discord bot.
Added a CLI.
Revamped logging to be more traceable.
Added IDs to Presences, to be used in the upcoming Transformation API.