Wrote a new frontend in Vue, made the renderer reusable within another Vue project, removed Discord command API, all pipes controllable from frontend, and more
APIError has a "fields" property now that can designate the fields that caused the error.
SecurityKit can generate a token for an existing user model rather than re-querying the database
Added logging configuration for the database
Reduced the number of hashes when generating crypto strings to improve API response time
Added safe optional chaining where it was missing
Improved sanitization of presence data before it is sent to the PresenceProvider from the AdapterSupervisor
Added RemoteClient-specific APIs for the frontend – many of these will not go through without an "identity" cookie, and will 401 for Authorization headers. This is because they often should only be run when a user has explicitly authenticated with their password.
The web framework now writes headers at the end of the response to prevent header collisions