Refactored the server package into some tinier packages that can run as independent microservices
To do this, a few things were required:
All security operations run in an isolated process with a REST API that is only exposed to other services. All transactions require either a password or a token, which, unlike an api key, cannot be generated without a password. This introduces an added level of security, and allows the encryption/decryption operations to not slow down other services.
Some database entities now live in shared-db, where database entities can be used across microservices. shared-db is also the host for ElasticSearch-indexed entities.
All redis code now lives in core-cache, where the cache helpers and generators for static structures live. This library handles reactivity to changes, and abstracts the redis and pub/sub backend.