Rewrote the server and module frameworks, switched to Redis for events. Removed a lot of bloat from the server package. Split additions up into two separate modules.
By removing the EventBus and switching to a more ipc-focused approach, modules in their new form can run across many processes. Modules can access the shared entities provided in shared-db, or they can access their own database and create entities on it. They have full access to the Presenti API.

The server package was so bloated and had a lot of legacy code, so I gutted it and rewrote it based on the new Server class in presenti/shared-infrastructure. This class serves as the foundation for any Presence Server instance, and will run with any configuration of modules. These modules all work together through Redis.

Normal web servers will not use the Server class, as they do not use the presence capabilities.

I'll continue reducing fragmentation and increasing the amount of shared code between projects, and will work towards a new docker configuration that can be deployed to production.

My next goals are to write specialized servers in Rust that integrate into the Redis event system.